requirements.txt
fire
GitPython
inflect
pandas
PyMovieDb
python-dotenv
requests
tqdm
twitchio[starlette]
discord.py
lxml[html_clean]